Separate bloc event and state definitions
