+++ lib/domain/bloc/sync_event.dart
@@ -1,0 +1,30 @@
+part of 'sync_bloc.dart';
+
+sealed class SyncEvent {}
+
+final class SyncGetTasksRequested extends SyncEvent {}
+
+final class SyncUpdateTasksRequested extends SyncEvent {
+  final List<Task> tasks;
+  final int revision;
+
+  SyncUpdateTasksRequested(this.tasks, this.revision);
+}
+
+final class SyncAddTaskRequested extends SyncEvent {
+  final Task task;
+
+  SyncAddTaskRequested(this.task);
+}
+
+final class SyncUpdateTaskRequested extends SyncEvent {
+  final Task task;
+
+  SyncUpdateTaskRequested(this.task);
+}
+
+final class SyncRemoveTaskRequested extends SyncEvent {
+  final String id;
+
+  SyncRemoveTaskRequested(this.id);
+}
